+++ src/listeners/panel/ArticleListener.ts
@@ -1,7 +1,6 @@
 import { Article } from '../../commands';
 import { ArticleHelper } from '../../helpers';
 import { PostMessageData } from '../../models';
-import { Command } from '../../panelWebView/Command';
 import { CommandToCode } from '../../panelWebView/CommandToCode';
 import { BaseListener } from './BaseListener';
 
@@ -18,7 +17,7 @@ export class ArticleListener extends BaseListener {
         Article.updateSlug();
         break;
       case CommandToCode.generateSlug:
-        this.generateSlug(msg.payload);
+        this.generateSlug(msg.command, msg.payload, msg.requestId);
         break;
       case CommandToCode.updateLastMod:
         Article.setLastModifiedDate();
@@ -33,11 +32,19 @@ export class ArticleListener extends BaseListener {
    * Generate a slug
    * @param title
    */
-  private static generateSlug({ title, slugTemplate }: { title: string; slugTemplate?: string }) {
+  private static generateSlug(
+    command: CommandToCode,
+    { title, slugTemplate }: { title: string; slugTemplate?: string },
+    requestId?: string
+  ) {
+    if (!command || !requestId) {
+      return;
+    }
+
     const article = ArticleHelper.getFrontMatterFromCurrentDocument();
     const slug = Article.generateSlug(title, article, slugTemplate);
     if (slug) {
-      this.sendMsg(Command.updatedSlug, slug);
+      this.sendRequest(command, requestId, slug);
     }
   }
 }